[execution] rename evm.{hpp,cpp} -> execute_message.{hpp,cpp}
The Yellow Paper defines EVM execution as several layers of state
transitions:

  - Block level execution, represented by the function capital pi;
    we call this `execute_block` (lives in execute_block.{hpp,cpp})

  - Transaction level execution, represented by the function capital
    upsilon; we call this `execute_transaction` (lives in
    execute_transaction.{hpp,cpp})

  - Message level execution of the messages that make up a transaction;
    there are potentially many of these in a single transaction, and they
    can be reverted; the naming around these tends suffer from confusion,
    and is (partially) corrected by this commit (more on this below)

  - Code execution; these are the arbitrary, user-programmable state
    transitions ("smart contracts"); this is often the component called
    "the EVM" and in our code, is implemented as separate library (found
    in category/vm)

The terminology around message execution tends to be confusing. There
are two kinds of messages:

  1. "Calls", represented by the YP function capital theta

  2. "Contract creations", represented by the YP function capital lambda

Both forms have the same input (e.g, `struct evmc_message` in the EVMC
API) and sometimes both forms end up being called "calls" even though
that is technically wrong. For example, `struct evmc_host_interface` only
has the `evmc_call_fn` callback. It takes a `struct evmc_message` which
can dispatch to either, and should probably have been called something
like `process_message` or `execute_message`.

Message processing has its own execution semantics: this is where the
balance transfers happen, nonces get incremented, frames get reverted
or committed, contract deployment addresses get calculated, etc. It is a
small but important, distinct phase of the Ethereum execution model. It
is primarily the place where the "most interesting" aspect of Ethereum
as a computation system happens: the atomic revert or commit of a frame.

For historical reasons, message processing was in a file called `evm.hpp`,
because the third (message) and fourth (code) execution layers were
lumped together. That is probably because the code execution layer was
provided by evmone and was not part of the project, so the naming
distiction was less important.

Note on terminology: both calls and contract creations create a separate
execution environment which can execute code, and which can be atomically
reverted; this object has no standard terminology, but is often called a
frame by Geth e.g., in the "call frame tracer."
