+++ category/execution/CMakeLists.txt
@@ -133,12 +133,12 @@ add_library(
   "ethereum/create_contract_address.hpp"
   "ethereum/dispatch_transaction.cpp"
   "ethereum/dispatch_transaction.hpp"
-  "ethereum/evm.cpp"
-  "ethereum/evm.hpp"
   "ethereum/evmc_host.cpp"
   "ethereum/evmc_host.hpp"
   "ethereum/execute_block.cpp"
   "ethereum/execute_block.hpp"
+  "ethereum/execute_message.cpp"
+  "ethereum/execute_message.hpp"
   "ethereum/execute_transaction.cpp"
   "ethereum/execute_transaction.hpp"
   "ethereum/fmt/event_trace_fmt.hpp"
